Add slidingWindowSum
`slidingWindow` typically helps write only functions with
complexity around `O(n*k)`, where `n` is the number of elements
in the stream and `k` is the size of the window. In many cases,
this can be reduced to `O(n)` by looking not at the window itself
but instead the sum of that window in some `Semigroup`. This can
be used, for example, to implement moving averages such as
arithmetic, geometric, or harmonic means.